% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixUploadedFilesNames}
\alias{fixUploadedFilesNames}
\title{Retain original file name of uploaded file}
\usage{
fixUploadedFilesNames(x)
}
\arguments{
\item{x}{result of a shiny fileInput call}
}
\value{
string with original file name
}
\description{
Retain original file name of uploaded file
}
\note{
see https://github.com/daattali/advanced-shiny/tree/master/upload-file-names
}
\examples{
\dontrun{
new_path <- fixUploadFileNames(input$my_upload)$datapath
}
}
